requirements.txt
anyio==3.5.0
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
executing==0.8.2
h11==0.12.0
httpcore==0.14.5
httpx==0.21.3
idna==3.3
ipykernel==6.7.0
ipython==8.0.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
notion-client==0.8.0
numpy==1.22.1
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
stack-data==0.1.4
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
